add some more code to method combinations specimen
It works now, but test case definition is far from perfect or even
remotely useful.
Points of interest: `call-method` and `make-method` (which is not yet
represented here) are special macros local to
`define-method-combination` form. Documentation on them available
online in the Hyperspec.

Replacing `call-next-method` with just `collide` inside :test
specified method causes Lisp to exhaust the stack (quite obviously).

Assert reports failure in a form of runtime error. So it may be useful
to replace it with some package-defined function to report cases
nicely.

Also if I proceed to use method testing as example, it would be nice
to add some convenience macros for defining test methods or even
define-method-with-test or something like this.